Trick IDEA into processing both transliterated and native forms
Without this implementation of "processNames" IDEA would never call
"processElementsWithNames" when the entered pattern does not have
a direct hit. For example: Let there be a sole file "Aenderung"
but no file that has a "Ä" as first letter. The uses enters "Ä" but
"Aenderung" would not show up. This is because the default implementation
of IDEA will only pass those file names to "processElementsWithName"
which already did match the search pattern in "processNames". But the
sole file "Aenderung" did not match "Ä" and hence is discarded.

Now both the actual name and its transliteration will be passed to the
default implementation of the name processor. One of the two will be
matched by the original processor ("Änderung" in this case although it
does not exist) and hence shows up in "processElementsWithName".
Finally that method looks again for both forms and finds the actually
existent matches.